Support storing objects bigger than 64kbs
Azure Tables has a column size limit of 64KBs. When storing unicode text, this is 32768 characters.
The AzureTables Storage provider stores the state object item as JSON.
The workaround is to store the object in chunks of 32768  characters, using the PartitionKey as the Key and RowKey as the chunk position.

- On Write: serialize to JSON and split into chunks, then create TableEntity with an incrementing RowKey. Delete any RowKey > than last generated RowKey (e.g.: cleanup chunks from a previous bigger object)
- On Read: Query the table for the specific PartitionKey, retrieving all Rows and ordering by RowKey
- On Delete: Query the table for the specific PartitionKey to delete, iterate over rows and delete them.